refactor: move the getBrowserExecutablePath to the common package
